+++ src/hooks/shortcuts/useVaultShortcutsHandlers.ts
